add precision to Fibonacci

+++ Algorithms/Fibonacci/FibonacciBinetsImplementation.cs
@@ -8,12 +8,12 @@ namespace Fibonacci
 {
     public static class FibonacciBinetsImplementation
     {
-        public static double Fibonacci(int input)
+        public static long Fibonacci(int input)
         {
             double phi = (1 + Math.Sqrt(5)) / 2;
             double psi = (1 - Math.Sqrt(5)) / 2;
 
-            return Math.Floor((Math.Pow(phi, input) - Math.Pow(psi, input)) / Math.Sqrt(5));
+            return (long)Math.Floor((Math.Pow(phi, input) - Math.Pow(psi, input)) / Math.Sqrt(5));
         }
     }
 }
